Fix proftool for profile file with 0 fill at end
Source: Mentor Graphics Corporation
MR: FBT-125
Type: Enhancement
Disposition: Needs submitting to upstream
Description:
The u-boot proftool expects the profile file to end exactly
on a record boundary. With writing mmc files, it is difficult
to get the profile file (written in 512 byte blocks) to meet this.

If the EOF occurs reading a header, proftool completes
normally, leading to inconsistent behavior. This fix causes
proftool to complete processing when it first encounters a
zero-filled header, or EOF.

Signed-off-by: John Mehaffey <mehaf@gedanken.com>